feature/implementing_mongo: mvoe standalone project to module server
