+++ airbnb-root/server/src/main/kotlin/ru/mai/dep810/airbnb/server/repository/MongoClientRepository.kt
@@ -1,11 +1,11 @@
-package ru.mai.dep810.airbnb.repository
+package ru.mai.dep810.airbnb.server.repository
 
 import org.springframework.data.mongodb.core.MongoOperations
 import org.springframework.data.mongodb.core.query.Criteria
 import org.springframework.data.mongodb.core.query.Query
 import org.springframework.stereotype.Repository
-import ru.mai.dep810.airbnb.data.*
 import java.util.*
+import ru.mai.dep810.airbnb.server.data.Client
 
 interface IClientRepository{
     fun findAll() : List<Client>
@@ -13,14 +13,15 @@ interface IClientRepository{
     fun deleteById(uuid: UUID)
 }
 @Repository("clientRepository")
-class ClientRepository(val mongoOperations: MongoOperations) : IClientRepository {
-    override fun findAll(): List<Client> =
-            mongoOperations.findAll(Client::class.java)
+class MongoClientRepository(val mongoOperations: MongoOperations) : IClientRepository {
+    override fun findAll(): List<Client> {
+        return mongoOperations.findAll(Client::class.java)
+    }
 
     override fun save(client: Client): Client =
             mongoOperations.save(client)
 
     override fun deleteById(uuid: UUID) {
-        mongoOperations.remove(Query.query(Criteria.where("_id").`is`(uuid)),Client::class.java)
+        mongoOperations.remove(Query.query(Criteria.where("_id").`is`(uuid)), Client::class.java)
     }
 }
