+++ airbnb-root/server/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+server.port=5000
+spring.data.mongodb.database=AirBnB
+spring.data.mongodb.host=localhost
+spring.data.mongodb.port=27017
+dump.rooms_path=data/rooms.csv
+dump.path=resources/Users.xml
+dump.needed=true
+elastic.port = 9200
