--- application.properties
@@ -1,9 +1,0 @@
-mongo.url=mongodb://localhost:27018
-mongo.database=tempr
-dump.rooms_path=data/rooms.csv
-dump.clients_path=data/clients.xml
-dump.load_needed=true
-spring.data.mongodb.host=localhost
-spring.data.mongodb.port=27017
-spring.data.mongodb.database=mongo-rest-api-demo
-elastic.port = 9200
