--- src/test/kotlin/ru/mai/dep810/twitter/TwitterApplicationTests.kt
@@ -1,3 +1,0 @@
-package ru.mai.dep810.twitter
-
-
